+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @casual-ui/react
 
+<<<<<<< HEAD
+=======
+## 0.0.14
+
+### Patch Changes
+
+- 122b697: - Bug: fix date picker useEffect loop
+  - Refator: reimplement the CTabs and CDatePicker with CCarousel
+  - Docs: Add type highlight for prop types
+- Updated dependencies [122b697]
+  - @casual-ui/react@0.0.14
+
+>>>>>>> c89a0a91ab350885c5c06e3c2b8d6fbfc904daae
 ## 0.0.13
 
 ### Patch Changes